requirements.txt
numpy
pandas==1.2.2
beautifulsoup4
python-binance
slackclient==1.0.7
dateparser==1.0.0
fake-useragent==0.1.11
tqdm